add deep equal check for tree rendering
