+++ package.json
@@ -98,6 +98,7 @@
     "q": "^1.5.1",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
+    "react-fast-compare": "^3.2.0",
     "react-hot-loader": "^4.13.0",
     "react-json-view": "^1.21.3",
     "react-redux": "^7.2.2",
